Add citations back.
This change creates a BibTex bib file and adds back citations
that were deleted from the document.

It also includes\ some housekeeping: add a .gitignore file to
igore files produced by LaTex.  Add a special command at the
top of every LaTex file so that TexWorks knows to build the
main file for the document when you tell it to run Latex
while editing the file.