Update the rest of the cross-references in the document.
The label names have the form the chapter:yyy or
section:yyy, where yyy is a short, unambiguous
descriptive string name for the chapter or section
with no spaces.

Names should be long enough to avoid ambiguity
when referring to concepts. For example,
function-call-argument is used instead
of function-argument because function-argument
could be ambiguously interpreted as referring to
a parameter variable.