Start learning to use TikZ to replace the first missing picture.
The conversion from Word to LaTex lost several relatively simple
pictures.  We'll replace them with pictures that will be
created using TikZ, a popular drawing package for LaTex.  For those
not familiar with LaTex drawing packages, you write a series of commands that
produce a picture within the LaTex document.

This check-in is a start on replacing the first picture, which is a
diagram of how relative alignment works given an object in memory and some
pointers to the object.

The alternative of using an external program to create pictures
and insert them into a LaTex document is not attractive.
The external program has to produce PDF (or Postscript) because
that's the scalable vector graphics formats that LaTex understands.
That is then inserted into the LaTex document. PDF from the Git
perspective is a binary file: there's no sensible human merging possible
for changes.  In addition, the visual results are usually poor anyway, with
mismatches in fonts, font sizes, and awkward scaling.  So we'll go
with the less visual approach of "programming" pictures.