Rewrite alternate possible semantics for bounds declarations.
This change rewrites the alternate possible semantics for bounds
declarations discussed in the design-alternatives.tex.  It
adds a discussion of eager vs. deferred evaluation for
bounds expressions.  It updates the discussion of when
bounds expressions are evaluated in variable-bounds.tex to
match this discussion.

It also discusses different possible definitions of the
extent of flow-sensitive bounds declarations.  The definition
of extent in variable-bounds.tex is changed to be simpler
to understand.  In particular, we outline all the conditions
that are required for a bounds declaration to apply to a
function component.  The prior description had a preceding
negative case (variables being out-of-scope) that caused
bounds(none) to apply instead.   The rewritten description
avoids the need for the reader to negate this condition
and add it to the other conditions to come up with all
required conditions.

This change removes the matrix of different orders of evaluation
(eager vs. deferred) and different definitions of flow-sensitive
extent.  It does not make sense to have the chart anymore
now that we have both lexically-scoped bounds declarations
and flow-sensitive bounds declaration.  The chart was an
attempt to find whether one combination of evaluation
order for bounds expressions with one definition of extent
for flow-sensitive bounds declarations produced a simpler
system.  Lexically-scoped bounds declarations make the
system simpler, and order of evaluation is largely orthogonal
to lexical scoping.