Add author section to the cover page.  Update fonts in macros.
This change adds a clear author section to the cover page for
the technical report.

It also updates macros for bounds expression and bounds declarations
to put their arguments in teletype font by default.  Arguments can have
their own fonts if desired.  This will cut down on the number of times that
we have to tell Latex to change fonts.  The \var macros already introduces
a different font for variables used with these macros.