Remove out-of-date comments about disjunction in bounds expressions
The introduction and open issues sections include
material about disjunction that is out-of-date.
We allow disjunction in non-modifying expressions.
- Update the material in the introduction to remove
the comments about disjunction not being allowed.
Also update the surrounding discussion to discuss
our approach to making static checking times
reasonable.
- Remove the material in the open issue section
about disjunction.  We are already allowing it.

Also improve the writing in the discussion of
design alternatives, and add an introduction
paragraph for the chapter.