Reorder design alternative sections to match order in main document body.
Move the section on allowing bounds expressions to be declared
with no relationship to the value of the expression to the
end of the chapter.  This matches the order in which topics
appear in the main document body (new types, operations on
those types, bounds declarations, and checking bounds declarations).