Add some highlighting of code back, fix formatting errors.
There was some code in the original Word document that was
highlighted to show changes.  Add the highlighting back
so that the changes are easily visible.  Also italicize
the code for printed copies that might be missing color.

Fix formatting of a BNF grammar.  One clause was broken
across two lines, which will cause confusion.  Another
clause was indentend when it shouldn't be.