Rewrite section on pointer cast operators.
The new approach changes the C cast operator
to have additional static checking to ensure that
casts to the different kinds of safe pointer types are
bounds-safe.  This extends the approach that was used for
the C cast operator for array_ptr values, where
the bounds propagate through the C cast operator
and the the bounds are checked for validity at the
statement level.

The core_bounds_cast operator is removed because
we are using the C cast operator by default.
dynamic_bounds_cast can then be built on top of the
C cast operator.   For assume_bounds_cast, we require
that the case where the upper and lower bound are
specified explicitly have relative bounds alignment
that can be proved at compile-time.

In all, this seems like a much simpler system to explain
and use.