Describe checking in unchecked blocks for expressions mixing safe and unsafe pointer types
This change describes the checking that is done in unchecked contexts
for expressions that mix safe and unsafe pointer types.  An example
of this is assignment of an expression with a safe pointer type
to a variable with an unsafe pointer type that has a bounds-safe
interface.   In this case, the declared bounds of the unsafe
pointer-typed variable will be checked at the end of the statement
containing the assignment.  This ensure that the safe pointer-typed
expression is being in a way that ia consistent with its bounds.

In general, checking will be done at the bounds-safe interface when
an expression with a safe pointer type is used where an expression
with an unsafe pointer type is expected.

This change also moves some examples to the section on implicit
conversions from unsafe pointer types to safe pointer types. It
updates the remaining examples and add some new ones.