Rework pointer/integer conversion description
Move the description of pointer/integer type conversions
and how to track bounds through integral expressions to
the interoperation chapter.  This material was in the
chapter on structures, but it seemed out of place there.
It was there because the example code used structure
types.  Rework the example code to not depend on
structures and just directly operate on tagged
pointers.