Checking validity of bounds when there are structures.
Update description of how to check validity of bounds
when there are structure types that have array_ptr
members.

* The material is short enough that there does
not need to be a separate chapter.  Fold the
material into the existing chapter on structures
and delete the separate chapter.
* Introduce the use of a representation that captures
the key idea of checking structure variables pointwise.
Structure variables are represented using the individual
variable member paths for array_ptrs.  The bounds
for values with structure types are represented using
individual member paths and the bounds for those
member paths.  With this representation, the rest
of the checking is straightforward to describe.