Rewrite material on pointers to arrayptrs.
Revamp the material on pointers to data that contains
array ptrs.  The prior write up had gone into the
weeds trying to deal with aliasing.  For pointers to
structures, what we really want to do is generalize
the treatment of structures from using variables
to using lvalue expressions, and use lvalue-producing
expressions in place of variables.  In practice, this
means  using expressions that produce lvalues that are
constant over some part of functions.  The discussion
now covers the 0th-order case, where we assume that
there is no additional alias information available.

Sketch out a preliminary proposal for dealing
with taking the addresses of variables used
in bounds declarations.

Add an acknowledgement section.