Update chapter on simple invariants
Remove redundant material about
facts and where clauses.  We modified
the original grammar for where clauses to
include facts.  We don't need to restate
it here.

Say that the facts about equality  of
variables and non-modifying  expressions
will be inferred automatically.  It
would be verbose to have to always call
them out.  Leave some details  about
transitivity of equality and inference
unspecified for now.