Testing parsing of new pointer types.
This test checks the parsing of the new Checked C pointer types, which are
available when the -fcheckedc-extension flag is passed.  The test checks that
the new types can be used as parameter types, return types, local variable
types, and structure member types.  It also checks the use of nested new
pointer types such as ptr<ptr<T>>, where >> could be mistaken for the
right-shift operator.