Test typechecking of expressions using new pointer types.
Add tests for typechecking the following kinds of expressions when new pointer
types are used:
 * Subscripting expressions
 * Assignments
 * Conditional expressions
 * Function calls

The tests have both positive cases (expected to typecheck) and negative cases
(expected to not typecheck). The tests are organized as sets of functions
for each kind of expression. There is one function that covers basic
functionality (non-void pointer types without cv modifiers) and then
additional functions to cover void pointer types and pointer types with cv
modifiers.  Within a function, tests are organized by feature, and within
features, tests are ordered by the kind of pointer type (unsafe, ptr, and
then array_ptr).

There are tests for various kinds of implicit conversions:
* Conversions between safe pointer types and unsafe pointer types.
  - Conversions from an unsafe pointer to a safe pointer type where
  the referent types are the same type are allowed.
  - Conversions where the referent types differ are generally not allowed,
    although there are special cases for void pointers.
  - Conversions between different kinds of safe pointers are generally
    not allowed.
* Conversions of pointer types to the boolean (_Bool) type.  These are allowed
  by the C11 standard for unsafe pointer types.  They are allowed also for safe
  pointer types.
* Conversions of 0 (the null pointer) to safe pointer types.
* Conversions between integer types and pointer types.  These are not allowed
  by the C99 or C11 standards.  Many compilers permit them and only issue
  warnings.  These are forbidden for safe pointer types.  The checking
  of bounds declarations would fail anyway, so it is better to issue a
  typechecking error first.
* Conversions involving void pointers.  void * can be converted
  to a safe pointer type at assignments and function calls.  Other void
  pointer types cannot be converted implicitly to non-void safe pointer
  types.  Safe pointer types can be be converted implicitly to a void pointer
  type that is the same kind of pointer.

  Note that conditional expressions are special: if one arm of a conditional
  expressions has a void type, the entire expression will have a void pointer
  type.  This means that the other arm will be implicitly converted to the
  void type.  This implicit conversion is not allowed for safe pointer
  types.

We avoid writing expressions with a subexpression that is expected to fail
to typecheck.  The type of the subexpression isn't defined, so there is not
well-defined behavior for typechecking the expression.  Compilers may
legitimately differ in whether or how they typecheck the expression.

We also try to avoiding writing expressions that will typecheck but
later cause an error during checking of bounds declarations.  Implicit
conversions of unsafe pointers to the ptr type generally must have known
bounds too.  We instead take the address of a local variable (when possible).