Implicit conversions and comparison operands.
Update the spec to reflect the fact that comparison operations do not
have implicit conversions for pointer operands.

Relational pointer comparisons (<, <=, >, >=) only allow pointers to objects
of the same type.  Equality and inequality comparisons allow comparisons
between pointers to objects  of the same type, as well pointers to void and 0.
Because the comparison operators are defined to allow values of different
pointer types, they do not do implicit conversions on wpointer type operands.

Add a few sentences to the definition of the comparison operators and
simplify the section on implicit conversions