Do not use safe and unsafe to describe pointer types.
We have been using the terminology safe and unsafe to describe the new
pointer types and the existing C pointer type, respectively.  The term
unsafe has negative connotations.  The term safe is also imprecise.  The
design only addresses bounds safety and does not yet address memory
lifetime safety or type casts that allow corruption of bounds information.
We are switching to the terms checked and unchecked instead.