Revise related work section in design document (#16)
This change converts the "lessons" section to a comprehensive related-work section. It includes an overview of the large amount of related work in the area, divided into 5 areas: (1) runtime-based approaches (2) security mitigations (3) static analysis (4) program verification and (5) programming languages. 

This addresses feedback from people that the design document needs to be improved to explain clearly how Checked C relates to other work in the area. This drops some of the detailed comparison with Deputy in favor of a broader overview of related work.