Add tests for type checking of checked arrays.
Many of the  tests are variants of the existing pointer type checking tests,
modified to use variables with array types instead of pointer types.   It also
includes some array-type specific tests:

* Tests that cover parameter variables with array types, where "array of T" is
retyped as "unchecked pointer to T" and "checked array of T" is retyped as
"array_ptr to T".
* Tests that cover pointers to arrays, including implicit conversions from
"pointer to array" to "pointer to checked array", where the pointers may
have different kinds.
* Tests for subscripting the initial dimension of multi-dimensional arrays.
These operations are interesting because they produce pointers to array types.

While writing these tests, I realized that we did not have tests for implicit
conversions between pointers to pointer types, where the outermost pointer had
a check or unchecked kind.  Conversions from unchecked to checked are OK,
but the reverse are not.  I added those tests to pointer_type_typechecking.c