PR 2126: Tests for unchecked keyword in array declarators.
This adds tests for using the unchecked keyword in an array declarator to declare an unchecked array.  This allows multi-dimensional arrays to be declared where accesses to the inner dimensions are not checked.  Note that 3 and 4-d arrays need to be used to test mixed checked and unchecked array dimensions, where checking propagates until it hits a nested array that is declared as unchecked.

Related work items: #13651