Add feature tests for parsing bounds declarations. (#33)
This change adds:
- Feature tests for parsing variable declarations with bounds
  declarations.  Most of the tests are for local variable declarations,
  with the assumption that parsing of bounds declarations themselves will
  be similar for variables with different linkage and storage classes.
  The tests also include tests for the other storage classes and linkage
  combinations that C allows.
- Feature tests for parsing function pointers where there are bounds on
  function parameters.

It also moves files to subdirectories so that tests will be easier to find.
The subdirectories are named after the functionality that is being tested
(parsing and typechecking, so far).   With the prior naming scheme, 
functionality was included at the end of the file name, making it harder to find 
tests.