Update description of checked pointers.
This finishes up addressing issue #11.  When we switched to the terminology of
checked and unchecked pointers, there was feedback that we should be very
clear about what is being checked.  Specifically, checked pointers check
bounds, but do not check memory lifetime.  This adds a paragraph to the
description of checked pointers making to clear that memory lifetime is
not checked.