expand implicit conversions for pointers to arrays (#41)
This addresses issue #20 by expanding the implicit conversions for pointers to arrays.  We now allow conversions from pointers to unchecked arrays to pointers to checked arrays.   This is needed to support passing unchecked multi-dimensional arrays to functions that have checked array parameters.

This change adds a notion of assignment compatibility.  Compatibility of types in C is similar to equality and is not directional.   We need a directional conversion, where source and destination matter.  Informally, we allow a type A to be converted to another type B if A and B are the same except that B has more checking.  

I considered defining a recursive notion of assignment compatibility for referent types that are pointers, but decided not to do that now.    It requires other changes to the specification that go beyond what is needed for issue #20, so I'm putting that work off.

I also reworked some of the examples for casts and conversions and added non-examples of code that will fail to check.