Add tests of parenthesizesd array declarators.
This change adds more feature tests of checked modifiers for parenthesized
array declarators.  These tests uncovered that the checked property
was not being propagated properly for parenthesized array declarators.
The tests also uncovered that the error message for a checked array of a
typedef that is an unchecked array was not correct.

Also shorten some variable names by removing the prefix 'single_'
and fix some indenting issues.