Test checked properties of dimensions of multi-dimensional arrays (#46)
This change adds feature tests that the dimensions of a multi-dimensional array are either all checked or unchecked.   The Checked C spec requires this for multi-dimensional arrays.     It removes feature tests where individual dimensions could be declared as unchecked using the `unchecked` keyword.   These feature tests were for an experimental feature in the clang implementation of Checked C.  We have decided to stick with the existing wording in the spec.

This change also adds tests for parenthesized array declarators.   The checked property is supposed to propagate from the outer dimension of a multi-dimensional array to the inner dimension: `int a checked[10][10]` declares a checked array of checked arrays  This propagation wasn't happening when the declarator for the outer dimension was parenthesized: `int (a checked[10])[10].`