Tests for checking type requirements for bounds declarations. (#48)
This change add tests for checking the type requirements of bounds declarations.  It matches a corresponding compiler change in  the checked-clang repo.

This commit include both positive tests (that are supposed to work) and negative tests that should fail with an error message. We have added the expected error message from the clang implementation for the negative tests, so that the tests can be used in the clang test infrastructure. This commit also fixes issues in existing tests that the checking uncovered.