Test type checking of bounds range expressions (#51)
This change adds tests for type checking bounds expression of the form bounds(e1, e2).  It tests both positive cases (that should type check) and negative cases (that should fail to type check).    This change matches a corresponding change in the checkedc-clang repo.  It addresses checkedc-clang repo issue 14.

It includes tests of cases such as using expressions whose types are given by typedef'ed type names, using an expression with an array type as a bound, and using pointers that point to the same object type but have different modifiers.