Change Checked C keywords in tests.
We are changing Checked C keywords to start with an underscore and a capital
letter.  This is clang Github pull request 53.  This is to avoid conflicts
with C identifiers in existing code.

This change adds a header file that #defines the existing keyword names to
the new versions of the keywords. It updates the tests to use the header file.