Change Checked C keywords to avoid conflicts with existing identifiers. (#52)
We're updating the clang implementation so that Checked C keywords start with an underscore and a capital letter. This avoids conflicts with existing C identifiers.

The keyword change follows the rules in the C Standard: all identifiers beginning with an underscore followed by a capital letter or underscore are reserved for implementation use.  This allows them to be used for language extensions without conflicting with existing identifiers. Of course, some programs do not follow the standard and use these identifiers.

This change adds a header file stdchecked.h that has macros that maps the existing non-prefix, lower case keywords to the new keywords.  It updates the tests to include that header file.

Testing:
- Updated version of clang passes these new tests.