Add tests of parsing interop type annotations. (#56)
This change adds tests for parsing of bounds-safe interface type annotations.   It matches a corresponding change for the checked-clang repo. 

The type annotations are a generalization of the `int * x : ptr` annotation described in the specification. That annotation does not provide a way to represent information for pointers with multiple levels of indirection (`int **x` for example). Instead of just using `ptr`, we now allow a type annotation to be specified.  The annotation has the form `type` `(` _type name_ `)`.   For example, a variable declaration `int **x` could be annotated as `int ** x : type(_Ptr<_Ptr<int>>)`.  This richer form allows us to represent bounds-safe interfaces where pointers are passed by reference, for example.

Bounds-safe interfaces are allowed for parameters, global variables, and members.   This change adds tests for type annotations involving `_Ptr`, `_Array_ptr`, typedef'ed names, and types with type modifiers.