Update error messages for return types that cannot have bounds declarations.
This change updates the error messages generated when a return bounds
declaration is see for a return type that cannot have a bound declaration.
For example, a function return a float cnanot have a return bounds declaration.

The existing error messages say that the return type has to be a pointer,
array, or integer type.  Functions cannot return arrays in C, so this
is confusing.  Update the error message to