Add rules to specification for functions without prototypes. (#66)
Update the specification to add some draft rules for declarations of
functions without prototypes.  These are C functions declared without their
arguments.  Informally,
- In checked scopes, it is an error to declare or use functions without
  prototypes.
- Outside of checked scopes, it is an error to pass a checked type or
  return a checked type from a function without a prototype.
The change adds a new section to interoperation that contains examples
and detailed rules.    The draft rules need to be reviewed with other people
working on Checked C.

The change removes functions without prototypes from the list of work
to be addressed in Chapter 10.2.  It also updates the section on checked
scopes to say that functions without prototypes are not allowed there.

This change updates some examples with empty parameter lists to
use void to properly describe the empty parameter lists.