Test redeclarations of no-prototype functions with checked types. (#67)
Calls to no-prototype functions are not type checked according to the C Specification (Section 6.5.2.2 of the C11 specification).  Because of this, the Checked C extension does not allow no-prototype functions to be redeclared to take or return checked types.   The lack of type checking at calls enables checking of bounds to be bypassed. Section 5.5 of the Checked C extension specification describes the restrictions.  In the terminology of the C Specification, no-prototype function types are incompatible with function types that take arguments or return values that have checked types, contain values with checked types embedded within them, or are recursively pointers to function types that take or return checked types or checked values.

This change adds feature tests that check that redeclarations of no-prototype functions to use checked types are not allowed.  The tests are in  typechecking/no_prototype_functions.c.

It also fixes accidental declarations in existing tests of no-prototype functions.  All these functions actually take no arguments.  In C, a prototype for a function that takes no arguments has to be declared with an argument list that is the `void` keyword.  For example, `int f(void);` declares a function that takes no arguments and returns an integer.