Change specification to describe updated Checked C keywords. (#68)
This change addresses issue #53.  To avoid backward-compatibility problems when using the
Checked C extension with existing C programs, we are changing the keywords for the
extension to start with an underscore followed by a capital letter.  

The updated keywords are:

_Array_ptr  _Checked  _Dynamic_check  _Ptr  _Span  _Where  _Unchecked 

This change adds a description of these keywords to the specification.  We use a header file 
that provides all-lowercase versions of the keywords.  The text and examples in the
specification continue to use the all-lowercase versions because that is the way extension is
intended to be used.

This change also clarifies that the syntax for bounds expressions uses specific identifiers,
not keywords.   This reduces the number of new keywords that need to be introduced.
The places where the specific identifiers occur in bounds expressions are not places where 
general-purpose identifiers can also appear.  This means there is no need to turn the specific
identifiers into keywords to disambiguate a possible overlap during parsing.  This technique
has been used in C++ and C# to avoid introducing new keywords for names that have special 
meaning only in specific syntactic contexts.