Rename some sections. (#71)
Update the title of Chapter 4 to be clearer.  The terminology
"check bounds" is ambiguous and could be misconstrued to be about
runtime bounds checks.  The chapter is really about checking bounds
declarations.

Update the title of "Extent of dataflow-sensitive bounds declarations"
to be "Extent of bounds declared at statements".  The terminology "dataflow"
is specific to compilers, so avoid it.