Add description of bounds-safe interface types. (#80)
This change reworks the bounds-safe interface section.  It adds a description of bounds-safe interface types and brings the section up-to-date with the clang implementation.  This addresses issue #60.

The text for the section needed to be revised substantially so that it accurately describes that programmers can use bounds declarations or bounds-safe interface types, depending on what needs to be described.  While I was doing that, I improved the writing and organization of the section.   I shortened up some descriptions.  I moved sections on type checking and checking of bounds declarations that are primarily for language designers and compiler writers to the end of the section.  I also added a description of the organization of the section.