Test redeclarations of functions with bounds declarations. (#83)
This change adds tests of redeclarations of functions with bounds declarations.  This  matches a corresponding change for the Checked C clang implementation.  The tests check that functions can be redeclared with identical bounds declarations on parameters and returns.  They also check that error messages are produced when the bounds declarations are not identical. This change also checks redeclarations of functions with bounds-safe interfaces. For now, the bounds declaration must be syntactically identical modulo parameter names.  Later this will be changed to require that the bounds declaration be provably semantically identical.

To further test redeclarations of functions, this change also adds header files that redeclare functions in the C standard library with bounds-safe interfaces, when the functions can have a bounds-safe interfaces.   This exercise shows three things, two of which were known and one that is new:
- There needs to be type system support for null-terminated arrays.
- That there is not a way to specify a bounds-safe interface to variable argument functions.
- That bounds declaration may need refer to parameters in an enclosing parameter list (bsearch and qsort demonstrate the need for this).   The Checked C clang implementation does not account for this.

This change also improves the testing of typechecking of bounds on parameters and return types, including bounds and returns of function types.