Update error messages for bounds mismatches. (#84)
The clang error message for bounds mismatches when functions were redeclared was "mismatched type", which was not helpful.   We've updated the Checked C version of clang to produce better error messages for mismatched bound.   This updates the error messages in the tests.   Also rename some functions from to begin with f instead of fn.