Deal with hiding of variables.
Add paragraph to the checking rules that requires variables  that hide other
variables be renamed before checking.  We used to require that variables
not hide other variables at the source level so that checking was sound.
We don't need this requirement at the source level and can deal with
hiding during checking by renaming variables appropriately.