Allow bounds declarations for integer-typed variables. (#90)
*Add a paragraph describing bounds declarations for integer-typed variables.   These are needed when a pointer is cast to an integer that is then assigned to a variable.  This brings the specification into agreement with the clang Checked C implementation, which already allows this.    This address issue #79.

Also fix some typos in the interoperation section and clarify some wording.
Adjust the tag in the example to only be one-bit wide.   There are some additional complications with handling multi-bit tags.   See GitHub issue #89 for details.  To avoid these complications, we just change the example to use a 1-bit tag.

Fix a bad LaTeX page break for an example by removing a line from an example,    There was a single line on one page with the bad page break.