Fixes non-portable casts in tests (#95)
The C Standard is purposefully vague on the subject of the size of pointers, and their comparison to the size of various integral types.

Here, we split the tests that contain integer <-> pointer casts into two groups. In the first group (the vast majority of platforms), `void*` is the same size as `long`, so we do all our casts to `long`s. In the second group (notably including Win64), `void*` is not the same size as `long`, but it is the same size as `long long`, so we do all our casts to that.

The only place this approach doesn't work is with Enums, which have an implementation-defined size (according to my reading of the spec), so I've commented out the pointer->enum casts. 

Fixes #92