Update tests to ensure automatic variables are initialized. (#96)
This change matches a corresponding change in the Checked C
clang repo.  The compiler will be checking that automatic variables with
_Ptr types or bounds declarations are always initialized using an
initializer.

This adds missing initializers to existing tests.  It also adds tests that
the compiler produces errors for uninitialized automatic variables.

Note that static variables are always initialized to 0, if they do not
have an initializer.  This means that we do not have to check that they
are initialized.  0 is valid for any bounds declaration, so it works well as
a default initialization value.