Adds Explicit Dynamic Check Tests (#99)
This adds the `dynamic_check` macro for `_Dynamic_check`, and some initial tests for its functionality

These initial tests make sure that successful programs don't exit with a failure on a dynamic check failure, and that dynamic checks that fail cause an exit with an error. Unfortunately, the execution of `llvm.trap()` counts as a crash for the "not" tool, which could make these tests more brittle. But they at least pass without having to add a SEH or similar.