Adds Verification of Explicit Dynamic Check Warnings (#106)
This updates some of the tests for Microsoft/checkedc-clang#128 to make sure we get some warnings on super-simple dynamic checks that will fail. 

We probably want to be careful to only use the `-verify` flag in a few tests, so the tests are not too coupled to clang's constant folder/default optimization level/the accuracy of our implemented static analysis. I think it's fine in these tests, but probably no others.