Fix casts of sources with unknown bounds to ptr types. (#115)
We are adding a check to the compiler that the source expressions of casts
to ptr types have bounds.  This uncovered a number of tests where the
source expressions do not have bounds.  Usually, the problem was a cast
from a variable with unchecked pointer type to T to ptr<T>.  The fix is to
use an expression that has known bounds instead, such as an expression
that takes the address of a variable with type T instead.