Use signals for out-of-bound memory access tests. (#123)
This way processes exit gracefully, instead of relying on the OS to handle
a crash.  On Windows, crashes cause Windows Error Reporting (WER) to fire up,
invoking system logging.

We have been using the not program from the LLVM test harness with the --crash
option to determine when processes have exited with a crash.  We have found
that this approach is not working reliably on Windows.  The not program spawns
a child process and looks at the process exit code, looking for negative exit
codes to indicate a crash. When running automated testing under Visual Studio
Team Services, sometimes the exit code is a positive number.

Testing:
- Changed tests run successfully for x86 Windows x64 Linux.