Renames parameters called 'ptr' in checked headers (#132)
Before this change, it was required that programmers were careful with the order in which they included headers - in particular, the inclusion of `<stdchecked.h>` must come after the final inclusion of `<stdlib_checked.h>` or the final inclusion of `<stdio_checked.h>`, whichever came later.

This is a problem if you want to be able to use the `ptr<T>` in a header file rather than `_Ptr<T>`, especially as this almost definitely brings the include of `<stdchecked.h>` earlier.

The solution is to rename these parameters. As these are declarations, all that matters is the types are compatible, the names are effectively erased (even if we use them in a `count(e)` or `bounds(e1, e2)` expression, these are canonicalised to argument position rather than name.