Remove span type from the language extension design.
The span type seems like a type that is better suited for a library than a core
language extension. In C++, which has templates, there is already a proposal
for adding a span-like type to the C++ Standard Library.  Remove the span
type for the language extension design and move it to the design alternatives
section.

To be clear, we are postponing adding the span type to the language extension
design (and the clang implementation as well).   We may decide to add it in
the future, especially because implicit conversions involving library types
and pointers can be messy.  However, for now,  it can wait until the need
becomes clearer.